(tui-lua) expose getxy
This was an annoying oversight that made the attr_id attr tag
useless as it is intended as a way for the client- side to annotate
types for faster lookups on mouse actions and so on. With the recent
implementation of annotated drawing, fetch + toggle attributes is
more important.

FossilOrigin-Name: e7cda6f2ce76b4e89609a254aa4f219949a0acb328a7ee645c5efb1bb7ba5997