(tui/lua/readline) add border-drawing helper
This adds arcan_tui_write_border to toggle the border attributes
correctly for a full (x1,y1,x2,y2) rectangle. Also add that as an
option (=default) to readline for the completion helper. The old
style can be reactivated by setting compact_completion option for
readline itself, though yet another indicator that we need some
env- packed arg to control look and feel for widgets.

Tui should assume privsep:ed from before so config file is not a
good idea, the COREOPT facility in SHMIF is still uncertain if it
should be cleaned up or otherwise dealt with. Exposing it to Tui
seems intuitively like a poor idea.