(egl-dri) streams -> atomic-explicit
Now that nvidia fake-wraps their old streams path in the GBM wrapper
and the code for that has propagated in enough blobdrivers we can move
from dealing with two halfassed and broken implementations to just one.

There are more impactful changes though: it is also time to slowly
deprecate the legacy mode entirely. Part of that is using atomic
modeset as the default, wire in explicit synch throughout (starting
with the modeset paths).

Paths likely to see regressions which require manual testing:

 - VT switching (like that isn't /roll already)
 - map_video_display_layer transitions from composited to dumb
   (durden: target/video/advanced/source_fs)
 - monitor hotplug

A very likely one that is tied to synch where we need fencing causes
jerky-rollback to a previous frame.

To get the old behaviour back:

    arcan_db add_appl_kv arcan video_device_legacy 1