(tui) further refactoring adventures
This moves tsm out of tui entirely and back into the terminal frameserver.
The copywindow widget is removed as are the labels that mapped to it since
the introduction of cross-screen blit should make it trivial to implement
in a cleaner way.

The ts debugging extensions are removed to cut down on layer violations.
Most tsm functionality should be back in order with some minor exceptions
due to flag management.

Next in this would be to provide a smaller state machine (e.g. st.c) as
a substitute for the rather bulky tsm, then drop the tui references to
hooks. With that all references to 'non-altscreen'/scrollback mode can
be evicted from tui.

FossilOrigin-Name: e0ad90b78d6afeca39b2332c63ace719c555589c4fdf0334dde36333c7ff7c7f