(audio) add stub platform
This quick hack makes it possible to build arcan with minimal support for audio.
The main purpose is to map out the choke points of the previous
refactor and move some things into the common engine in order to also
make the platform parts smaller before adding support for other
backends.