(tui) add dangling 'swallow' subwindow req.
This lets a new window hint that it should take the place of its
parent until end of life. It is still a hint, the WM scheme can still
route inputs and other events to the parent and the application should
respond properly still.

this was 'fixed' quite a while ago but for what can only be thought
of as git misuse the changelog entry got in but not the actual
implementation.