(core) shareglstore, src->dst mixup for nullsrc
The behaviour for image_sharestorage was wrong for non-textured
src sharing into dst. Sharing an incomplete or non-textured store
should transform dst into the equivalent of a null surface, and
instead it had that effect on the src (changing from null to null).
Sharing a colour src should transfer colour and shader to dst.