(tui-lua/nbio) ref race when ok/bad in pollset
This one took a while. The tui_process result bitmaps were forwarded
into the various forms of data callbacks, one being the data_handler
function that may rearm or disable further callbacks. If that happened
on a descriptor with a good status while there were also data on a
bad one pending, the rearming could cause the wrong lua reference to
be rebound to the data handler, causing the next registry lookup to
return the wrong time, triggering the pcall/recovery chain. The
debug hooks and indirections added to figure this out are kept around
should there be similar dragons lurking somewhere.