(shmif) permit 'audio' only segments
This adds a new segment type for audio dominant data. The reasons for
this are threefold - in edge cases where there are ongoing audio streams
and subject to dynamic resizes, sharing a segment with the main video is
subject to underruns or forcing higher latency by prefilling with larger
audio buffers to work around the blocking nature of resizes over a
certain threshold. Similarly, many audio libraries prefer to have an
audio only thread, and the added synchronisation to work with the event
loop and resize behaviour of the parent was problematic.

The second reason is to carry better audio processing pipeline metadata
(mainly FFTs) that tend to be available at the source end from sampling
or resampling activities, and not repeatedly recreate them in near every
stage of an audio processing pipeline. For an audio only segment the
video buffer carries null- data or the corresponding FFT. This allows
the server end to pair such a transfer when there is new audio to handle
and avoid emitting it as a video frame but rather a resource that can be
accessed with the corresponding audio functions - allowing both an on
CPU and on GPU access.

The third reason is for surround sound and beamforming, instead of
having a matrix of many n.m.o formats multiplexed - the more
game/vr/xr/compression friendly approach of dynamically creating
positional channels and deferring the HRTF mixing for later stages.