(ttf)  cache repeated use of fonts
Glyph sharing in arcan_TTF (derived from SDL_TTF) between multiple consumers. The upside is that this hits all font use, both appl and TUI consumers. The downside is that we drift further away from resynching the changes made in SDL_TTF over the last few years, some of them quite useful. Accepted/merged for the time being, with the note that it might be sufficient to have the caching on the agp_vstore level when atlases gets used and if so, revert to upstream SDL_TTF.

FossilOrigin-Name: 65193befc612d8819656c1aa7f326b27498ad74897d4190830276bf2b6058e1e