(decode) h264 and mjpeg in uvc capture
This adds support for the missing two major formats for camera capture.
The weights for automatic format selection now favours h264 over mjpeg
over various raw ones. The next step here is extending shmif for video
passthrough. Even though this normally makes little sense for
afsrv_decode as its job is to be the [compressed formats] sink, the
situation is a bit more nuanced when it is combined with arcan-net.

For that scenario it is a substantial gain being able to maintain
compression from source, network transport to final sink and thus a good
test case for these kinds of setups.