(shmif/net) allow h264 passthrough negotiation
This allows a shmif_ext resize call to try and apply venc for setting
annotated encoded video frames. This changes the interpretation of the
supplied video buffer to be a preencoded fourcc tagged video source
along with framesize annotation.

The main use for this is with arcan-net where it can allow a h264
source, e.g. a webcam over hvc that expose such frames, to allow its
frames to go through to the other end unscathed.