(shmif/lua) add target anchorhint
This is a concession to primarily allow toolkits and protocol bridges
to systems that expect information about not just visibility and desired
dimensions but also position and hierarchy.

A side-effect to this is that it should make it rather trivial to put
together a miniwm appl (or durden workspace owner tool) that just
informs and delegates window management feedback to an external process
written in whatever.