(shmif/lua) add target anchorhint
This is a concession to primarily allow toolkits and protocol bridges
to systems that expect information about not just visibility and desired
dimensions but also position and hierarchy.

A side-effect to this is that it should make it rather trivial to put
together a miniwm appl (or durden workspace owner tool) that just
informs and delegates window management feedback to an external process
written in whatever.

FossilOrigin-Name: 489081163f0a0039dcfcaffcf1681da7b2715ad6a664253b28bfe5b17d1d1d7b