(shmif/decode) enable h264 passthrough in uvc
This follows the previous passthrough commit with its implementation
in uvc-support for decode. Shmif-server is also modified slightly to
revert to uncompressed frame if the fourcc gets unset for a frame in
order to support transitioning back without an explicit synch pass.

My camera is acting up so fairly untested still and the _RESET event
is still not respected meaning that the harder test case:

  1. hook up through afsrv_decode (no passthrough)
  2. force-redirect to arcan-net -s test (while camera active)
  3. reset-camera feed and swap to passthrough.
  4. redirect-back and disable passthrough.

is still not handled.

FossilOrigin-Name: 65da5f7279f70c9e097c774816aa647f65f8ce826e2231f724676263cbaecc91