(lua/shmif/egl-dri) wire up hdr metadata
This will version bump shmif as some things go from deprecated to
removed. The HDR16 metadata _ext substructure gets fleshed out and
renamed across all layers.

Some minor steps are missing still before durden et al. can get
away with tonemapping and full colourspace controls, but this should
be the main patch that takes changing interfaces.

On the Lua level video_displaymode with the argument format that
takes a table can now add keys for controlling the coordinates of
targetted display primaries and whitepoint, as well as update the
mastering and targetted light levels.

Next patch will cover the actual drm toggles for activating the
metadata, as well as providing the added information to the frame
delivery event so that the WM can decide to update display metadata
similarly to how the CM subprotocol works.