(shmif) mousestate helper improvements
Previously the caller was always responsible for maintaining the
storage for the state blob which was a poor decision. As to not break
API we overload and accept a NULL mstate holder to mean 'use the
internal shmif tracking structure'.

FossilOrigin-Name: 17dce2cea96d31823794375550c8734c4dd9a89fbe5373bf359adc0fb4dea811