(core/platform/lua) support keymap forward
In order to deal with top-down and recursive handling of clients
that are stuck with xkb forever (so x11 / wayland) it would help
to not sideband the keymap in at the bridge level but rather get
it through the engine so it would work network transparently.

This patch adds xkb to all platforms (should it be desired and
available) and adds the wiring to extract the current map or the
spec of one from lua via input_remap_translation, and then pass
it onward through open_nonblock(vid, ...).

This match changes done to durden as well as to Xarcan, arcan-wayland
bits missing still.