(core/lua/shmif) wire up image_metadata
This should provide most of the missing pieces for ensuring both a
path from a shmif consumer signalling HDR capability, annotating
frame with metadata and having that persist across the stack.

What is mainly missing for testing this for real is accessing the
metadata from Lua space (only really needed if you manually need
to composite / tone-map different metadata- sourced images in the
same pipeline.

The rest would be to simply add the properties to the atomic
modeset when a vstore with metadata set is mapped to an output.

FossilOrigin-Name: ad823fde4dd0efe5a31f1005e38109c191e870bc82f61ff3f9bcff65d57228ab