(shmif) merge compatible stepframes, preroll fdfix
With the clockreq changes from b3364d the chance for a buildup on
a stalled client is substantially higher. While realtime performance
is to be asked of someone explicitly asking for vblank feedback, it
is not always in their control. To help congestion then, add stepframe
to the list of valid candidates.

This also fixes a subtle thing to take note of and monitor for. Normally
fd- transfer events aren't necessary during preroll, with a notable
exception being DEVICEHINT for specifically routing a GPU, though that
has not yet been used. While re-using shmif for other purposes, it was
discovered that the first descriptor transfer during a preroll could be
dropped, never to find the descriptor. Whether this is posix-madness
that will come back to bite us or not is to be seen, but change the
behaviour to retry when an expected descriptor does not arrive - instead
of returning an event with a bad descriptor field.