(a12) allow clients to push appl update
This allows clients with the correct permission groups (default is deny)
to update and (soon) add new appls to the directory index. These are
currently not persisted, and there's some reason to not permit that
before the feature is more reliable and as an option.

It also wires up SIGUSR1 on the directory server process to force a
rescan/rebuild of the index (discarding any user-provided updates).

The intent for that is to have an accepted set of shared appls, allow
devs to test out new ones before committing to it. That commit can come
as a sideband git push into a trigger that sends the SIGUSR1. The
fancier option would be to move this into a database store of its own,
but walk before flying and all that.

With the appl- push also comes the scaffolding for having server side
appl shared resources, with two levels of updates. One with the style of
a traditional 'form upload' - while the other a more privileged
corresponding to appltemp file creation.

In the same vein, things are now mostly prepped for attaching a
'controller' to a hosted appl, meaning server side scripts to act as a
message router / server side logic for networked appls.

The last big hurdle in this is to rework how the client side launches
arcan-lwa, and specifically the monitoring interface. Currently that is
just a block FILE* hack with some plaintext strings, but we kindof need
background processing to support hot reloading, and with that there's no
point to not go full shmifsrv and gain the advantages of having the
control / monitor channel also speak shmif in lwa.