(a12) refactor dircl to handle bg processing
This changes the dircl runner around to keep processing a12 packets
while running an appl. While the monitoring channel still is primitive
it should be enough for moving on with live-reloading appls on update.

The refactor also allows for running multiple appls simultaneously over
the same a12 connection, although it is not permitted / enabled yet as
it makes other parts of the stack more difficult to develop right now.

It should also be a much smaller step bootstrapping shmifsrv over the
monitoring channel and using that to better integrate the appl with the
a12 connection. This one is important as the last hurdle before adding
the controller layer to the directory server, more extensive state synch
and appl-appl messaging.