(core) conductor workaround for _lwa
This is not a pleasant solution, but it gets to take one more for the
team until all the tracing and refactoring can be weighed in into proper
performance models / heuristics as part of the performance push.

Basically _lwa has an issue with display clocking as its external
'clock' is STEPFRAME events. These come as part of the regular input
event processing and is thus done when there is rendering in flight
and so on.

In comparison, egl-dri knows which parts is locked as the conductor
deadline is set with the target output in mind and the processing
pipeline is designed for this pattern.

Normally when there is load the _lwa use works as you get a natural
cadence from presentation -> interaction (with possibly animations
further causing a consistent load) but when it is idle and/or transition
to/from an idle state is when the headache starts.

This patch adds the main display for LWA (which is also a frameserver)
as an event sink with the subsystem normally used for alt_nbio. These
were biased to schedule in between frames as to interfere as little as
possible with the normal pipeline without adding the complexity of being
both asynch and threaded in a process with the complexities of GPU
compoisition in mind.

It also feeds in the deadline told to the conductor by the frame
delivery mechanism so that the pollset processing isn't overextending.

Synch is the real enemy.

FossilOrigin-Name: 50b0140b0ee09920717b57d48349dea175076050dff565188f886f8ea6e6837f