(a12) allow hot-appl reload
This should stitch things together from HEAD~2,3. By extending the
monitoring mode and watchdog handling arcan-net can now interrupt
the chainloaded arcan process both while busy in lua-vm state and
elsewhere in the processing.

This adds a reload command to the monitoring mode, as well as unpacking
an update appl into a .new folder. When the unpack is completed,
chainloaded arcan jumps into monitoring mode, -net switches applname.new
over to just applname and proceeds to tell arcan to reload the current
app. This triggers a behaviour similar to system_collapse(self) with the
regular adopt etc. procedure.

There are more variants to handle with this - e.g. allowing the user
to block this behaviour, to handle revert should the new appl be buggy
and so on but this should be enough to test the general behaviour in a
more networked setting.

FossilOrigin-Name: 1e1e48aeeed4d31177844afcdaff46407e980676a9930950d5113d4ea1608cfa