(a12/shmif) re-order pledge/unveil, add minimalfd
This fixes the privsep setup for directory workers on openBSD,
it also adds a profile that only takes stdio+fdpassing which makes
a single shmif connection work, but cannot crash-recover, resize
or create new windows. While not always ideal it fits with the
workers for directory server and for most afsrv_decode settings.