(a12) directory mode support for dynamic sources
This adds all the back and forth needed to coordinate a sink requesting
to access a source attached to a directory. With this comes most of the
boilerplate needed for both nat punching and STUN as well.

In rough parts the dance goes like this:

{client-source}: register(source, name)
{server}: broadcast to all workers
{server-worker-n}: if listening, send name
{client-sink-listening}: get new source, check and a12(diropen, kpub)
{server-worker-sink}: netstate(kpub) -> {server}
{server}: netstate(kpub) find worker with kpub -> gen. secret, msg
source/sink with ip info
{server-worker-sink}: a12(diropened)
{server-worker-source}: a12(diropened)
{client-sink}: connect(diropened_rep, secret)
{client-source}: (diropened:expected-kpub,secret}
                 -> (fork+rexec-self)
                 -> listen(secret, kpub}
{client-source}: on connection: authenticate (expected-kpub, secret)

After this there is a direct connection between the source and the sink.
There is a lot of robustness work still and the handover to listen still
does not function.

FossilOrigin-Name: 8584abf8f2a5515167dd7b7a3ad593d2c83080dada12ae61c6d9e32d14905a25