(a12) directory source-listen working
This patch adds all the necessary plumbing to negotiate a connection
between source and sink where the sink and directory is on a public
facing IP and listens for an inbound connection.

While a lot of testing remains, particularly cleaning up timeouts
and error handling - the hard work for this feature is done with.

Next up would be to do the same for tunneled connections where the
dir_srv_worker reserves a tunnel-ID, bchunks source and sink together
with sockets that copy directly between the workers, with two commands
added to the protocol - one for saying how many bytes should be written
in the tunnel before resuming normal a12 processing, and one where
the same goes for reading.

Thereafter we have nat-punch and sink listen left for all the basic
communications. One part that is still in question is whether it should
be allowed for the source to specify / override the ephemeral key for
the outer connection. The reason for that would be to allow the source
to expose itself with one identity to the directory and another to
the sink.