(a12) warning nits, src-port ctrl, prep-tunnel
Drop some dead code that didn't pan out, add controls for the listening
port when sourcing so that it's possible to test / run on the same
machine. Prepare for directory server tunnelling.

The plan for this is to simply reserve a transfer channel similar to how
bstream is already used, but without a bounded transfer length.

It'll require another thread in both dir_srv_worker and dir_cl to handle
the bidirectionality requirement, as well as a two specialised a12.h
functions, one for writing into the state machine and one for setting
the target descriptor for the inbound data.

When those work on a test- 'AAAA' stream, it should just be a matter
of having an extended form of the -S argument for the forked subchild
that substitutes inherited socket numbers for the matching socketpair()