(a12) initial directory-tunnel working
With this patch the end for tunnel juggling is in sight. This lands
the full path from a client requesting that a source be opened in
tunnel-mode. The server main process splits out a socket pair and sends
to the workers. The workers adds that to the i/o multiplexing and
reserves channel-1 for direct forwarding by piggybacking on bstream-
transfers.

The sink end spawns off a thread that latches into a tunnel pair and
feeds that into a socket that authenticates and maps to a regular
a12-cl-to-shmif connection, with corresponding feed into the tunnel
side of that a12 state machine.

The source end double-forks off (* this does perform some of the illegal
from-fork actions) and performs a similar dance, though no added
threading.

A12_IDENT=test ARCAN_CONNPATH=a12://mydirsrv afsrv_terminal
arcan-net --tunnel mydirsrv "*test"

should be roughly what is needed assuming an up to date and compliant
directory server.