(a12) change appl package format
This moves away from exec:ing to tar into a placeholder custom. The need
for this is to have a .manifest file for an appl that turns into a
global header for the appl with basic information about sandboxing needs
(primarily of -net and -terminal should be permitted or not).

Then there's a single line header per file, the raw bytes and repeat
until eof. Checksumming and compression is not included as the protocol
does that already. What is missing in that regard is signature chains
and caching of compression to reduce server spinup times, but that's
for the optimisation stage.

This should also restore functionality for afsrv_net browse/load that
was broken during the net/dir_cl refactor. The biggest discrepancy
between the two right now is that no NETSTATE events are emitted yet
and no tunnel request controls.

FossilOrigin-Name: e52827eb2ae32ac41710f13a6e6f5aa060bbd58f071f9328951c9a16046d201d