(a12) don't enable guard on openbsd
This is sandbox specific so should be extended as other platforms
get the same treatment. Basically disable the guard thread that
adds a little bit of protection against a crashed server for the
workers (marginally useful here but a little, not versus the
tradeoff of having the extra syscalls needed to manage the threads).

FossilOrigin-Name: 6b92fdbb0691a8734924530669d1c03d4aef2e77431fb5ead5b1ea9eb63fba9e