(core) don't force adopt handler ack on ep-create
Fun one. The flow for appl switching with adoption as well as with crash
recovery first runs the application entry point then the adopt handler.

This is deliberate so that window/resource management code doesn't
suffer from initialization order problems. The side effect from that is
that if a frameserver connection is established in the entry point, it
too will be exposed in the adoption handler. This has mostly been
relevant for durden, but it has had a deferred / reopen on termination
scheme that masked the issue.

When testing a12- based hot network reloading with the @stdin bit, the
net_open call does create a frameserver for the monitoring parent to
connect through in order for shmif event mapping to be less painful.
This had the side effect of getting pruned immediately through the
missing _adopt handler, uncovering the issue.

The solution as it stands is to track which reset- counter a frameserver
was created and registered in. If they are the same it means that it was
actually created in the main entry point and the script can be assumed
having proper tracking for it.

FossilOrigin-Name: 809f2480366adc21bb05ecaf65e56b42b54259dfc0275c3300dddf5b0011ec4a