(core) dirty state propagation test
This is a non-fix but useful regardless as we get slightly better tracing info.
The full issue is still that if an inner rendertarget has an attachment that is
dependent on a vstore in an outer rendertarget, and that vstore gets updated as
part of a frameserver poll sequence pre-render of the outer rendertarget, and
there is nothing going on in the inner rendertarget it will not update until
some other event inadvertently triggers it. This gets more obvious at low
refreshrates.

FossilOrigin-Name: cdf54231146ad5e31b4a67d3e5bcd2f7cf1a7b17493cd1b762185c32a3a76de8