(shmif/tui) add fd to local path resolver and map to on_bchunk in tui-lua
This is a necessary evil to avoid extraneous bgcopy actions when handling an
incoming descriptor. On supported OSes it tries to match a descriptor to a
local file path and matching inode. This makes it possible for the recipient
to forego creating a local copy first in order to pass to a consumer which
requires access to a filepath.

FossilOrigin-Name: 892094c89a8a9e9585f7af74e2a70a496cea3147cb03d91b9d0c8d85a95b5d5a