(lua/tui-lua) fix leftover 'number' breakting tui-lua error reporting
This was an evil one. It turnes out that the bit.c we vendored left a probe
number at the top of the stack not accounted for when exposing the tui bindings.
This meant that the pcall which uses an absolute index to reference the error
catching function that can derive the backtrace hits this number rather than
the proper function, being the major cause for why C -> Lua -> C -> Lua chains
were hard to decipher.

FossilOrigin-Name: aa870a64dd7c09ac43beb7d12b8a61f68a5882ed7fe994b9ab653a4c472f80d0