(shmif) wire in ACCESSIBLITY defimpl hooks and allocate a primary slot
This follows the pattern of the push-DEBUG segment to automatically build an
internally managed attached segment for providing some kind of support should
the client itself not provide an implementation.

The current form simply maps it, converts into TUI and writes back that there
is no 'alt' text. The next logical step is to use it and shmifsrv to spawn an
afsrv_encode process latched to related-source window video buffer submission
and use that with the tesseract_ocr mode or with a future LLM one to provide
content description.

FossilOrigin-Name: c01fceef633c2e616c2561ea2497bd57d20ebc66f0f005bc5e77bcb178957dc3