(core/lua) wire in text_surface
This should become a substantial optimization for the likes of durden.
render_text is overly complex as it contains embedding of images, etc. which
forces it to be treated as a full texture buffer for the time being.

More often than not though, the properties desired when rendering have beeen
closer to that of a regular arcan-tui window with minor changes in font and
style versus mixing and matching per video object.

The actual definition format is still in flux, but this wires in the basic
call and allocation, especially the gritty details of building a tui context,
matching it to the default font and current display density etc.

It is still rough and only partially tested, the main contenders would be
to make fonthints apply as well as trigger rerasterization if the primary
display attachment density changes.

FossilOrigin-Name: e1342b9538c77813be3e54d29f6903da9d24185e485cee046d2538b507f63912