(core/shmif/decode) add 'spell' proto stub and RHINT_EMPTY
This prepares for using a shared external spell-checking oracle with hunspell.
It will be fairly trivial as it only needs to support MESSAGE with the word to
get completion for, and a series of MESSAGEs back, terminated with a VSIGNAL.

The decode frameserver should be started with verbose frame reporting to
retrieve that frame signal, so that it is possible to distinguish between
requests properly.

To reduce the GPU transfers that would happen on SIGVID, a rhint is added that
lets the frameserver transfer skip texture store updates yet still emit the
correct events up the stack.

FossilOrigin-Name: 174272805ab8da4f89bf7ae2ae3717e19505cd1da59e1218be193e50fa900694