(core) don't default mark as dirty when there is a fract timestamp shader
This is a complicated remnant from the dirty tracking done before we allowed
direct rendertarget mapping to an output sink. The consequence was that if
there is a shader requesting fractional timestamp updates in an otherwise
static scene, the worldrt or related rendertargets would be skipped but we
still reported that there were updates to the platform layer.

This caused lwa to swap active rendertargets when there was nothing actually
drawn, effectively forwarding 'old' frames. The consequences of this change
may be that static single-object 'animations in shader' would only update on
ticks. The proper fix then would be to check for each rendertarget if there are
active visible objects with shaders that have fractional updates and only then
process it.

FossilOrigin-Name: 006451e16d7d1b38fd4c8ab1ef64090a2f236099f279f4c9658a4f69879cbbed