(net) workaround for confused interpretation of STATEDIR in resolving keystore root
ARCAN_STATEPATH is used to pass the root for all state store as far as arcan is
concerned, and this is used with all frameserver creation. The keystore subsequently
reserved the 'a12' subdirectory for use as a keystore as to not clash with anything
else stored there.

Directory mode using a config- file on the other hand would treat the STATEPATH as
being the keystore root as it is set explicitly through the lua. This clashed when
trying to beacon broadcast the keystore that is also used to seed the directory.

This commit adds a workaround for that by checking for the keydirs in root first,
and if they don't exist, follow the 'a12' format.

FossilOrigin-Name: 2a24e0903265ca41195bcb779d1ae85120df3bbcb8161301fe3a8e86f14a79f9