(net) rework how 'remember' interactive works
There are a few nuances to this. By default the trust_domain option
("outbound") is used to remember an unknown key. If the user specifies a custom
tag, it would not be triggered when making a new connection as it only
explicitly checks the trust-domain.

To remedy this, we also register the key in the hostkeys part of the store, so:

arcan-net arcan.divergent-desktop.org ->
unknown key, do you trust?
remember -> dd

arcan-net dd@

would work on successive calls. This has the caveat that there is no
differentiation for the key, the 'default' one used for outbound will be kept
compared to how it works when you manually run:

arcan-net keystore dd arcan.divergent-desktop.org

which explicitly generates a new outbound key for use with only that tag.

FossilOrigin-Name: 10bf5a2e86d6faea0696e76c1a2ea78ae14e6ace6b6785237805c7eee77702a5