(net) move file ul/dl handling from _dir to a12.c
This will break the current implementation, troubleshooting pending. The longer
story is that the workarounds used in _dir to handle submitting transfers were
hacky and can be cleaner merged inside of the regular a12 handling by matching
to the corresponding SHMIF events.

What this first round does is to check on enqueue if any special management is
needed to account for the other end being a directory. If so, the proper
BCHUNKSTATE event is tied to the pending transfer, and submitted before the
bstream header gets sent.

This should give better queue handling for uploads, but is incomplete for
download and download notification still.

FossilOrigin-Name: 7b61c86944d8ba1a42d27e308c993dcd95c78bd3fa3cbcdf20871326e51e4694