(lua) change alt_call to convey a trigger name
This is intended for automatic breakpoints on entrypoints into the Lua VM.
This defines a large bitmap of possible tracesources that we then map to
upcoming breakpoint command in the monitor interface.

That would then set the mask of active triggers, and when the mask matches,
change the lua trace function to break on the first instruction in the
entrypoint.

FossilOrigin-Name: 5aa80a095c8176c41612128bd060872d8195b84c419202b27b3e32745e212026