(lua) open_nonblock should give precedence to UD on mismatch
This is an API design oversight that lead to the confusion around BCHUNK_IN/OUT.
Take the case:

   open_nonblock(netvid, false, "somefile") -> ud_in
	 decvid = launch_decode()
	 open_nonblock(decvid, true, "*", ud_in)

For routing a file from net to decode. On the net-side we receive a BCHUNK_IN
that gets converted to a BCHUNKSTATE request and the server would enqueue the
bstream on the channel. We receive that and convert to a BCHUNK_IN that gets
bound to ud_in in read mode.

Now the last call is conflicting, it would result in ud_in descriptor
transferring to decvid but decvid would get BCHUNK_OUT. Since we can't break
the prototype for this function, ignore the direction request and use whatever
matches ud_in.

FossilOrigin-Name: 740bba9438f60dd769ee0e4d8a5670c4c1d7e1f9c5a1512619c9ec0509eec248