(net) start on passive-synch
Passive discovery just prints out the known tags discovered but if the IP is
different there is no way to synch to the keystore as of yet. This is
problematic as there can be many hosts to one tag, and each has (potentially) a
different keypair. This means that in order to update the keystore we would
have to probe every item in the tag until we find one that authenticates, and
then update the IP, if one is even used.

If it's host-based and the external naming system is out of date (DNS TTL ..)
changing it may have adverse effects so it should only apply to hardcoded IPs
(which LAN discovery was meant for anyway since the tag is a naming system).

FossilOrigin-Name: 82017f571b6021cb70c4be65d2cae736bbba0cdce5dab393c7cd7e15ef6dcd8e